modularize metadata container
